+++ 03-frontend/angular-ecommerce/src/app/components/product-list/product-list.component.ts
@@ -17,7 +17,7 @@ export class ProductListComponent implements OnInit {
 
   // properties for pagination
   thePageNumber: number = 1;
-  thePageSize: number = 10;
+  thePageSize: number = 5;
   theTotalElements: number = 0;
   
 
@@ -109,4 +109,11 @@ export class ProductListComponent implements OnInit {
     };
   }
 
+  // updates page size based on user choice 
+  updatePageSize(pageSize: number) {
+    this.thePageSize = pageSize;
+    this.thePageNumber = 1;
+    // to refresh view
+    this.listProducts();
+  }
 }